requirements.txt
websockets==12.0
keyboard==0.13.5
zeroconf==0.131.0
pywin32==306
pillow==10.1.0
pystray==0.19.4